win9x: sweep EncodedJSValue empty-return fix to 15 more bindings files win9x: fix libuv ssize_t ABI mismatch on 32-bit (ReturnCodeI64) win9x: document subprocess-spawn ssize_t + GQCSEx stub fixes win9x: document the vendor-patches/<name>/ auto-discovery mechanism
- The JSVALUE32_64 return {} bug (raw 0 decodes as JS double 0.0 instead
  of the EmptyValueTag) also affected EncodedJSValue-returning functions in
  other bindings files. Apply the same JSC::JSValue::encode(JSC::JSValue())
  fix across the files verified (via clang AST dump + direct brace parsing) to
  contain EncodedJSValue-returning functions: BunObject, BunString, ErrorCode
  (76 sites), H2HeadersMaterializer, IPC, JSBuffer, JSBunRequest,
  JSBundlerPlugin, JSCommonJSExtensions, JSDOMExceptionHandling,
  JSPropertyIterator, JSWrappingFunction, NodeVM, UtilInspect, ZigGlobalObject.

- A separate pre-existing libuv win32 IOCP teardown abort
  (GetQueuedCompletionStatusEx: (109) broken pipe) can still fire when the
  process shuts down after using a subprocess pipe; spawn itself now works.

- ReturnCodeI64 (libuv ssize_t) width on 32-bit: uv_read_cb nread and
  uv_fs_t.result are intptr_t (4 bytes on win32), were declared as 8-byte
  i64 -> garbage upper bits -> PosOverflow panic on piped subprocess reads.
  Now wraps isize.

- xp_compat.cpp GetQueuedCompletionStatusEx stub: dropped failed OVERLAPPED
  completions (e.g. ERROR_BROKEN_PIPE on a subprocess pipe), leaving
  GetLastError()=109 -> libuv uv_fatal_error abort at shutdown. Only
  WAIT_TIMEOUT means nothing was dequeued; failed ops are now delivered.

- vendor-patches/<name>/*.patch are auto-discovered by the build
  (source.ts emitFetch) and applied for every profile building the dep from
  source, on top of the explicit patches: list in the dep's .ts.

+++ src/io/PipeReader.rs
@@ -1435,7 +1435,7 @@ impl WindowsBufferedReader {
                 // of 74468 bytes). Just ignore 0-byte reads and let libuv continue.
                 return;
             }
-            v if v == uv::UV_EOF as i64 => {
+            v if v == uv::UV_EOF as isize => {
                 let _ = this.stop_reading();
                 // EOF (buf is not safe to access here)
                 return this.on_read(sys::Result::Ok(0), &mut [], ReadState::Eof);
@@ -1471,7 +1471,7 @@ impl WindowsBufferedReader {
         // `&mut WindowsBufferedReader` (distinct heap allocations).
         let (file, result, parent_ptr) = unsafe { crate::source::File::from_fs_callback(fs) };
         let nread_int = result.int();
-        let was_canceled = nread_int == uv::UV_ECANCELED as i64;
+        let was_canceled = nread_int == uv::UV_ECANCELED as isize;
 
         bun_sys::syslog!(
             "onFileRead({}) = {}",
@@ -1522,12 +1522,12 @@ impl WindowsBufferedReader {
 
         match nread_int {
             // 0 actually means EOF too
-            v if v == 0 || v == uv::UV_EOF as i64 => {
+            v if v == 0 || v == uv::UV_EOF as isize => {
                 this.flags.insert(WindowsFlags::IS_PAUSED);
                 this.on_read(sys::Result::Ok(0), &mut [], ReadState::Eof);
             }
             // UV_ECANCELED needs to be on the top so we avoid UAF
-            v if v == uv::UV_ECANCELED as i64 => unreachable!(),
+            v if v == uv::UV_ECANCELED as isize => unreachable!(),
             _ => {
                 if let Some(err) = result.to_error(sys::Tag::read) {
                     this.flags.insert(WindowsFlags::IS_PAUSED);
